MASSIVE commit attempting to clean up some of the CVS branches that have gone wayward.

+++ modules/inboxmodule/actions/delete_list.php
@@ -3,6 +3,7 @@
 ##################################################
 #
 # Copyright (c) 2004-2005 James Hunt and the OIC Group, Inc.
+# All Changes as of 6/1/05 Copyright 2005 James Hunt
 #
 # This file is part of Exponent
 #
@@ -31,22 +32,27 @@
 # $Id$
 ##################################################
 
-if (!defined("PATHOS")) exit("");
+if (!defined('PATHOS')) exit('');
 
-$list = null;
-if (isset($_GET['id'])) {
-	$list = $db->selectObject("inbox_contactlist","id=".$_GET['id']);
-} else {
-	echo SITE_404_HTML;
-}
-
-if ($list) {
-	if ($user && $list->owner == $user->id) {
-		$db->delete("inbox_contactlist_member","list_id=".$list->id);
-		$db->delete("inbox_contactlist","id=".$list->id);
-		
-		pathos_flow_redirect();
+if ($user) {
+	
+	$list = null;
+	if (isset($_GET['id'])) {
+		$list = $db->selectObject('inbox_contactlist','id='.intval($_GET['id']));
+	} else {
+		echo SITE_404_HTML;
 	}
+	
+	if ($list) {
+		if ($user && $list->owner == $user->id) {
+			$db->delete('inbox_contactlist_member','list_id='.$list->id);
+			$db->delete('inbox_contactlist','id='.$list->id);
+			
+			pathos_flow_redirect();
+		}
+	}
+} else {
+	echo SITE_403_HTML;
 }
 
 ?>
